% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suicide.R
\docType{package}
\name{suicide}
\alias{suicide}
\alias{suicide-package}
\alias{suicide}
\alias{batch_suicide}
\title{Let a process kill itself}
\usage{
suicide(interval)

batch_suicide(interval)
}
\arguments{
\item{interval}{number of seconds to sleep}
}
\description{
Let a process kill itself
}
